requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-bootstrap3==10.0.1
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-el-pagination==3.0.1
django-js-asset==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.5
django-simple-blog==0.2.1
django-simple-math-captcha==1.0.7
gunicorn==19.9.0
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.4.5
psycopg2==2.7.5
whitenoise==4.0
docx-mailmerge==0.4.0
django-mathfilters==0.4.0
